requirements.txt
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
bs4==0.0.1
certifi==2022.9.24
lxml==4.9.1
requests==2.28.1
pycryptodome==3.15.0

